rename build 'orchestrator' to 'strategy'
