+++ source/Nuke.Common/CI/ChainedConfigurationAttributeBase.cs
@@ -7,7 +7,7 @@ using System.Collections.Generic;
 using System.Linq;
 using JetBrains.Annotations;
 using Nuke.Common.Execution;
-using Nuke.Common.Execution.Orchestration.Sequential;
+using Nuke.Common.Execution.Strategies.Sequential;
 using Nuke.Common.Utilities.Collections;
 
 namespace Nuke.Common.CI
